EDIT orchard hostname to filter out http://
The new auto devops parameter service.url passes in a full url with
"http://" in front of it. When we use this url for the host field, it
fails the subdomain validation regex due to the "://" characters not
being allowed in subdomains. So we just have to remove the "http://"
from the url for use in the hosts field.