EDIT statspace faster find with large collections
When there are many collections, the API call for getting the
collections start to get slow. There's an option to enable an
"optimizedFindAuthorized" that speeds this up if enabled. I've set it to
the default "false" just in case it messes things up in StatSpace and
BioSpace. Only CWSEISpace needs this optimization.