ADD orchard deploy api
The UI seems to have deployed fine, so now deploying the API. There's
two main differences with the API, it has several environment variables
and it has a rewrite ingress.

The API uses environment variables to get the database info and the
token encryption info.

In order to avoid CORS errors and workarounds, I used a reverse proxy in
docker-compose to make it look like the API and UI were on the same
server. I've attempted to recreate this with the ingress settings, but
don't know if they'll work. The way I'm hoping it works is that all
traffic to /api/ gets sent to the API server while all other traffic
will be handled by the UI server.