FIX orchard UI js/css being blocked by browsers
Turns out all requests to the UI were loading index.html. So where the
browser is expecting javascript or css, it got html and rejected it. The
path rewrite wasn't working for the UI because the regex capture group
didn't exist for the UI. Changed the path regex capture group instead to
make it all work.