FIX alchemy review pods crashing
The review pods are probably crashing because the init-db-and-migrate
container doesn't depend on wait-for-db-service so doesn't actually wait
for the database to come up. I think it works on prod/staging because
the database in those cases startup faster due to persistence. The
review deployments don't have persistence so the database always has to
do the fresh instance initialization.

I've just removed the init-db-and-migrate service for now. It did
actually used to init the database, but I've moved that into the
entrypoint.sh script on the container itself. The only thing is does
nowadays is just add an admin user and we don't need that for
prod/staging nowadays. It would probably be better to move the admin
user creation to the entrypoint.sh script too.

I'm not sure if the wait-for-db-service actually works. But the image
we're using has its own wait for db script too since we need it for the
docker-compose version to work. We could probably remove the
wait-for-db-service but I don't want to yet until I've got some time to
troubleshoot in case things go wrong.