FIX axlrd fail deploy due to pv race conditon
The node deployment uses a persistent volume (pv). The rbd pv only
allows one pod to mount it at a time (for writes). So there's a rare
condition where it fails to bring down the old pod in time during an
upgrade, which stops the new pod from deploying due to not being able to
mount the pv. At least, that's the theory. So now we're forcing the node
deployment to use the Recreate strategy for deployment so that it'll
kill all existing pods before bringing up new ones. Hopefully it works.