EDIT lti-shim remove configmap, ADD sanctum confg
The configmap generates a Laravel .env file. This turns out not be
necessary as environment variables can be used to pass the same values.
So I've removed the configmap and started passing all the configuration
values as environment variables.

With the addition of Laravel Sanctum, two configuration values must be
configured correctly:

* SESSION_DRIVER must be set to cookie, it was previously set to
database
* SANCTUM_STATEFUL_DOMAINS must be set the domain this is deployed on,
session cookies need to match this domain in order to be considered
valid