FIX lti-shim deployment not updating pod
I initially thought this was because of the RollingUpdate deployment
strategy, so tried switching it to the Recreate strategy. This didn't
work.

But now it looks like it's because the generate deployment.yaml had no
changes and looked like the same file. Found some posts suggesting that
we add a date to deployment.yaml so that there is a changing dynamic
value that makes helm trigger a pod restart. Not sure where this dynamic
value should be. Tried putting it in the deployment metadata and it
didn't work, however, putting it in the template metadata label worked.

Note that naming the field "timestamp" results in it not showing up in
the deployment. So switching back to naming it "date".

I think this wasn't an issue before because we were using the
--recreate-pod or something similar (--force?) in the old gitlab-ci.