FIX lti-shim ingress apiVersion dep warning, typos
Apparently, the networking.k8s.io/v1beta1 API versions of Ingress will
no longer be served in v1.22. Looks like the serviceName and servicePort
fields were renamed as the major changes. A new 'pathType' needs to be
specified, if I'm reading this correctly, the value "Prefix" should be
identical to what was happening before.

https://kubernetes.io/docs/reference/using-api/deprecation-guide/

Fixed some typos for the mail deployment.