EDIT orchard define image in values.yaml
Define the image used in values.yaml. This is due to gitlab autodevops
passing values incompatible with docker hub in image.repository
(lthub/orchard/master when we only want lthub/orchard).

Autoformatter keeps complaining about the spacing of
spec.containers.name, so I just let it autoformat it to get rid of the
warnings.