EDIT lti-shim restrict mail access to internal ips
loadBalancerSourceRanges is a new field in Kubernetes 1.3. I'm not sure
if our kubernetes instance supports it, so trying it out. Version wise
it should, but apparently it might pass syntax validation but the load
balancer might not actually support it.

loadBalancerSourceRanges limits which client IPs can access this
service. The mail relay servers are all internal IPs, so this should
work. This should also allow all traffic that comes from within the
kubernetes cluster itself.