UPDATE celery invocation for celery 4 to 5 upgrade
The biggest breaking celery change is to the cli option order. Global
options, such as --app, needs to be placed after the celery command and
cannot be placed after sub commands (like worker). So I've fixed the
celery invocations I've found. Note that --loglevel needs to be after
the sub command, it's unrecognized as a global option apparently.