ADD alchemy cronjob for redis persistence rewrite
The redis persistence file needs to be rewritten from scratch
periodically, otherwise it'll just keep growing in size due to its
append-only nature. This is usually done via cron, but since the redis
image itself doesn't have cron in it, we need to do it via kubernetes
cronjob.

This is scheduled monthly.