FIX statspace 0.5.1 chart is missing
Turns out this is due to ingress api version deprecation.

Both 'extensions/v1beta1' and 'networking.k8s.io/v1beta1' are
deprecated. Helm throws a error if you try to use them. I've migrated
over to 'networking.k8s.io/v1'.

This doesn't really fix the missing 0.5.1 chart, as I incremented the
chart version to 0.6.0, but it'll let gitlab CI deploy this again since
it'll grab the newer chart.

Another helm lint error popped up where it didn't like that the app
version in Chart.yaml was a float and wanted a string, so I've just
added a semantic version number which is treated as a string.