EDIT lti-shim add readinessProbe to all deployment
Gitlab AutoDevops deploy keeps failing. I think it's cause none of the
pods report ready, they only report running. Guessing that previously,
'running' status was good enough to consider the deployment a success.
So trying to see if the pods reporting 'ready' will fix this.

Note that even though the pods only report 'ready', they're actually all
running fine and serving traffic correctly. It's just that the
deployment gets reverted by timeout when the pods don't report ready.

All the timeouts check every minute to see if the pod is serving
traffic. For the ltijs-mongo pod though, it's technically only seeing if
the mongo ping is working, not that mongo is capable of serving traffic,
but I didn't want anything more complicated.