FIX alchemy busybox nslookup broken, use wait4x
nslookup is used to wait for the database to be up. Latest busybox has a
bug in nslookup that causes this to fail. Fix is to switch to wait4x,
this is just copied from the compair fix for the same thing.

Helm lint didn't like the appVersion and app.nginx.configuration being
undefined, so I fixed those too. It also didn't like that service.url is
undefined, but that seems like it should come from gitlab, so didn't fix
that.