EDIT mediawiki new logo config for mw1.39
The old $wgLogo config is deprecated. The new $wgLogos config allows you
define many logo variations. Most of it is for backward compatibility.
The new mediawikiLogo.icon value is required as it is the only one used
by new themes from now on, all the other values will be deprecated in
the future.

Since we still have the old theme available as an option for users to
pick, we'll still use some of the backward compat logo values:

mediawikiLogo.legacy1x: for the old 1x versions of the logo
mediawikiLogo.legacy2x: for the old 2x versions of the logo

I'm not sure if both is required, but I got odd results when only one or
the other is present, so putting both in to be safe. We only have the 2x
version of the logo previously, so will have to upload a 1x version.