EDIT ipeer web proxy fastcgi_read_timeout config
The web proxy was timing out in 1 minute and some iPeer requests needed
longer to process, so I've added a values config web.timeout that lets
you configure fastcgi_read_timeout to a longer time.

I suspect our kubernetes ingress is already on a longer timeout, as I
didn't need to explicitly configure proxy-read-timeout and such for
ingress for the longer timeout to work.