EDIT webwork for 2.18 upgrade
New worker, the mojo worker. I've just copied the old worker section and
renamed it. Now we have two worker sections in values, one is
worker.lti1p3 and the other is worker.mojo

The new webwork container serves on port 8080 instead of the old 80. I'm
kinda surprised we only needed to change this in one spot but since it
seems to work, not complaining.