EDIT webwork support for saml2 conf file
Added another file for the configmap, previously we only had
localOverrides.conf. I renamed the "localOverrides.conf" key to
"localOverrides" as the dot was causing troubles, couldn't figure out
how to escape the dot and this was the simpler solution.

There is now a 'authen_saml2' file under .Values.webworkFiles, this is a
YAML file and needs to be mounted in Webwork's conf/authen_saml2.yml in
order to enable the SAML2 plugin.

Helm lint wasn't happy with the "if .Values.webworkFiles.localOverrides"
check, complaining about nil pointer. Found people saying that if you
wrap (.Values.webworkFiles) in parens, it stops complaining.