ADD chart for responsemap
Skeleton generated by helm. Ported over from the original Kubernetes
yaml files (old enough to be before we started using charts!).

Responsemap is a standard PHP app, served using Apache, with a MariaDB
database.

I considered using the Bitnami MariaDB chart, but decided it was really
overkill for a tiny app like this.

The PVC can only be read by one pod at a time, so the update strategy is
to recreate the pod instead of being rolling. Otherwise the new pod gets
stuck waiting for the old pod to release the PVC mount.

Had to mount the upload directory as www-data or image uploads won't
work.