ADD mediawiki support for simplesamlphp extension
This involves 3 main changes:
- Add simplesamlphp deployment
- Route paths starting with /_saml2/ to the simplesamlphp container
- Mount simplesamlphp code in both the simplesamlphp container and the
  main wiki container

The simplesamlphp deployment is basically running simplesamlphp in SP
mode. The SP certs are provided in a configmap and mounted into the SP.
The certs had to be mounted individually as they were being mounted on
top of a pvc volume, don't know why mounting as a dir doesn't work in
such cases.

The simplesamlphp SP needs to be hosted on the same domain as wiki to
avoid cookie issues, so ingress was configured to serve all traffics
with paths starting with /_saml2/ to the SP.

On the wiki side, the simplesamlphp extension needs to import classes
from the SP codebase. This is why the code for the simplesamlphp app
lives on a shared volume. When the SP starts up, its entrypoint copies
the code into the shared volume, and apache runs the SP off of the
shared volume code. The wiki mount is set to readonly.

Incremented the minor version in Chart.yaml as this is a larger change
than usual.