ADD cardscanner helm chart
This adds 3 pods:
- php-fpm app server that runs the actual cardscanner laravel site
- nginx web server that serves static content and passes other requests
  to php-fpm.
- laravel worker that runs possibly large jobs separately (currently
  just LTI nrps syncs)

The nginx web server needs access to the static assets in order to serve
them. This is accomplished by mounting shared directories read only in
the nginx pod. The public/ shared directory shares compiled css/js and
is rsynced by the php-fpm pod on each bootup. The storage/ shared
directory is written to directly by php-fpm.

The shared directories are NFS mounts and might contribute to some
slowness, not sure this is the best solution, but it works for now.

If it weren't for the shared directories, I think ingress could be
configured to forward requests to php-fpm. I couldn't figure how to
serve static assets from shared directories wtih ingress.

The readiness and liveness probes for php-fpm uses fcgi-probe to make
requests directly to php-fpm. The /up health check is a laravel feature.

The nginx probes likewise make requests to /up. So if something happens
to php-fpm, the nginx pod might also get rebooted.

The probes are very forgiving in terms of timeout right now, they should
probably be tightened later. I wanted to set a long window before it
errors out for manual troubleshooting purposes on the initial deploys.