FIX css/js links using http, autodevop deploy fail
Laravel doesn't properly generate https urls cause it doesn't know it's
serving behind a load balancer that's taking care of https. I've set
nginx to add the X-Forwarded-For header and for Laravel to trust that
header. Hopefully that fixes it.

Autodevops deploy fails at the last point when it tries to check on the
rollout status. This is because it uses the release name to check if our
main deployment is successful, but _helper.tpl defines the deployment
name as release name + chart name. I've changed it to only use release
name as autodevops expects, hopefully that fixes it.