ADD cardscanner lti config as env vars
Pass in LTI config from values file as env vars to the main php-fpm app
and the worker. The worker needs the LTI settings since it does the LTI
syncs.

The LTI_DEPLOYMENT_ID can take multiple deployment IDs, enter as a space
delimited string.

LTI_PUBLIC_KEY and LTI_PRIVATE_KEY is expected to be a base64 encoded
PEM. Added comments to the values file for how to properly generate
values for these two fields.